# 🚀 Phase 1 Performance Optimization - Fixes #28