## 🚨 URGENT PERFORMANCE FIX

This PR addresses the critical performance issue described in #28 where the application was causing high CPU usage and fan speed even when running locally.

## 📊 Performance Improvements

### Major Optimizations:
- **Entity Count**: Reduced from 40 to 15 per symbol (120 → 45 total entities)
- **Frame Rate**: Reduced from 60 FPS to 30 FPS with proper frame limiting
- **Collision Detection**: Implemented spatial partitioning for O(n²) → O(n) performance
- **UI Updates**: Throttled from every frame to every 100ms
- **Visual Effects**: Disabled particles, trails, and celebrations for performance
- **Grid Drawing**: Disabled for better rendering performance

### Audio System Fixes:
- Fixed sound file paths to use correct `assets/sounds/` structure
- Optimized audio loading and playback
- Reduced console spam from audio errors
- Improved audio context initialization

### Technical Optimizations:
- Spatial grid updates throttled to every 50ms
- AI behavior updates reduced to every other frame
- Visual updates reduced to every other frame
- Removed unnecessary audio cloning
- Increased collision grid size for better performance

## 🎯 Expected Results

- **50-70% reduction in CPU usage**
- **Maintained all core game functionality**
- **Improved battery life on mobile devices**
- **Reduced fan speed on local development**
- **Fixed audio system functionality**

## ✅ Testing

- [x] Server starts successfully
- [x] Game loads and runs without errors
- [x] All core functionality maintained
- [x] Audio system working
- [x] No linting errors

## 🔧 Files Modified

- `docs/js/config/game-config.js` - Reduced entity count and FPS
- `docs/js/game/game-engine.js` - Frame rate limiting and grid optimization
- `docs/js/utils/audio-manager.js` - Fixed paths and optimized playback
- `docs/js/app.js` - Throttled UI updates
- `docs/js/game/collision-system.js` - Spatial partitioning optimization
- `docs/js/entities/base-entity.js` - Reduced update frequency

Closes #28